últimas implementações das views - Desafio concluído
Adicionado:
- modais par informar a conclusão das operações executadas.
- implementação da logica para carregar dinamicamente a pagina de edição dos dados dos veículos.
- Implementação da logica para salvar as alterações feitas nos dados dos veículos.
- Implementação da logica de exclusão de um veículo.
Alterações:
- Métodos da controller implementados para retornar mensagens de erro quando ocorrer.
- Nome da variável de objeto Moto alterado para evitar inconsistência na conversão para objeto JSON.

+++ gestaoFrota/src/main/java/com/sergipetec/gestaoFrota/controller/VeiculoController.java
@@ -6,6 +6,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Map;
 
 @RestController
 @RequestMapping("/veiculos")
@@ -17,34 +18,54 @@ public class VeiculoController {
     }
 
     @PostMapping
-    public ResponseEntity<Veiculo> cadastrarVeiculo(@RequestBody Veiculo veiculo) {
-        Veiculo savedVeiculo = service.cadastrarVeiculo(veiculo);
-        return ResponseEntity.ok(savedVeiculo);
+    public ResponseEntity<?> cadastrarVeiculo(@RequestBody Veiculo veiculo) {
+        Veiculo savedVeiculo = null;
+        try {
+            savedVeiculo = service.cadastrarVeiculo(veiculo);
+            return ResponseEntity.ok(savedVeiculo);
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().body(Map.of("message", e.getMessage()));
+        }
+
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Veiculo> atualizarVeiculo(@PathVariable Long id, @RequestBody Veiculo veiculo) {
+    public ResponseEntity<?> atualizarVeiculo(@PathVariable Long id, @RequestBody Veiculo veiculo) {
         veiculo.setId(id);
-        Veiculo updatedveiculo = service.atualizarVeiculo(veiculo);
-        return ResponseEntity.ok(updatedveiculo);
+        Veiculo updatedVeiculo = null;
+        try {
+            updatedVeiculo = service.atualizarVeiculo(veiculo);
+            return ResponseEntity.ok(updatedVeiculo);
+        } catch (Exception e) {
+            return ResponseEntity.badRequest().body(Map.of("message", e.getMessage()));
+        }
     }
 
     @DeleteMapping ("/{id}")
     public ResponseEntity<Void> excluirVeiculo(@PathVariable Long id) {
-        service.removerVeiculo(id);
-        return ResponseEntity.noContent().build();
+        if (service.DetailsVeiculo(id) != null) {
+            service.removerVeiculo(id);
+            return ResponseEntity.noContent().build();
+        }
+        return ResponseEntity.notFound().build();
     }
 
     @GetMapping
     public ResponseEntity<List<Veiculo>> listarVeiculos() {
         List<Veiculo> veiculos = service.listarVeiculos();
-        return ResponseEntity.ok(veiculos);
+        if (veiculos != null) {
+            return ResponseEntity.ok(veiculos);
+        }
+        return ResponseEntity.notFound().build();
     }
 
     @GetMapping("/{id}")
     public ResponseEntity<Veiculo> getById(@PathVariable Long id) {
         Veiculo veiculo = service.DetailsVeiculo(id);
-        return ResponseEntity.ok(veiculo);
+        if (veiculo != null) {
+            return ResponseEntity.ok(veiculo);
+        }
+        return ResponseEntity.notFound().build();
     }
 
 }
